+++ HSharp_v0.4_Tests/tokens.py
@@ -25,6 +25,8 @@ class TokenType(Enum):
     MODULE = 'MODULE'
     CONCEPT = 'CONCEPT'
     CORO = 'CORO'
+    ASYNC = 'ASYNC'
+    AWAIT = 'AWAIT'
     ASM = 'ASM'
     PTR = 'PTR'
     TRY = 'TRY'